validasi delete history

+++ app/src/main/java/com/muchlis/inventaris/views/fragment/CctvHistoryFragment.kt
@@ -80,7 +80,9 @@ class CctvHistoryFragment : Fragment() {
         bd.rvDetailComputerHistory.layoutManager =
             LinearLayoutManager(requireActivity(), LinearLayoutManager.VERTICAL, false)
         historyAdapter = HistoryAdapter(requireActivity(), historyData) {
-            deleteComputerHistory(it.id)
+            if (it.author == App.prefs.nameSave){
+                deleteComputerHistory(it.id)
+            }
         }
         bd.rvDetailComputerHistory.adapter = historyAdapter
         bd.rvDetailComputerHistory.setHasFixedSize(true)
